pyproject.toml
[tool.poetry]
name = "bhumi-bot"
version = "0.1.0"
description = ""
authors = ["pavel <pshishkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.1"
motor = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
